requirements.txt
arcade==3.3.2
esper==3.4
pytest==8.4.1
pyyaml
psutil
tqdm==4.67.1
numba==0.61.2